feat: wip implement new ProjectItem component with dialog and carousel integration
